+++ core/src/main/java/org/eqasim/core/simulation/EqasimConfigurator.java
@@ -9,6 +9,7 @@ import org.eqasim.core.components.transit.EqasimTransitModule;
 import org.eqasim.core.components.transit.EqasimTransitQSimModule;
 import org.eqasim.core.components.transit.routing.DefaultEnrichedTransitRoute;
 import org.eqasim.core.components.transit.routing.DefaultEnrichedTransitRouteFactory;
+import org.eqasim.core.simulation.calibration.CalibrationConfigGroup;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.api.core.v01.population.Person;
@@ -28,8 +29,8 @@ public class EqasimConfigurator {
 		return new ConfigGroup[] { //
 				new SwissRailRaptorConfigGroup(), //
 				new EqasimConfigGroup(), //
-				new DiscreteModeChoiceConfigGroup() //
-		};
+				new DiscreteModeChoiceConfigGroup(), //
+				new CalibrationConfigGroup() };
 	}
 
 	static public List<AbstractModule> getModules() {