+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 - Add iml and idea to gitignore
+- Add calibration utilities
+- Improve code structure of the choice models in `core` and `switzerland`
+- Improve trips analysis with PersonFilter and give access to other modules if present
 - Add analysis of trips via config file
 - Add better support for command line parameters
 - Bugfix: Utility calculation for car/pt did not consider cost